feat(home): live-network globe fed by real server locations
Replace the "voices" globe (invented quotes pinned to censored capitals)
with a DB-driven live-network globe:

- Green markers are real server locations from publicConfig.locations
  (operator-set coords on backendServers, the same feed as the status
  strip and the account globe), each with a label chip showing code,
  city, and online state; offline instances go gray.
- The amber censored-region dots stay as static, unlabeled context; a
  legend under the globe says what both colors mean.
- CobeGlobe.svelte -> NetworkGlobe.svelte; the quote scheduler, per-id
  CSS visibility rules, and line-rotation clock are gone (chip
  visibility is JS-driven off the same front-face + collision pass).
- The section hides until at least one location has coordinates, so the
  "real network, shown live" claim never renders empty.
- home.globe.* catalog copy rewritten in English; stale locale
  translations of the old section removed so they fall back to English
  pending native review (home.globe.voices deleted everywhere).
