+++ convex/_generated/api.d.ts
@@ -34,8 +34,6 @@ import type * as keyRevocations from "../keyRevocations.js";
 import type * as lib_accountId from "../lib/accountId.js";
 import type * as lib_analyticsConfig from "../lib/analyticsConfig.js";
 import type * as lib_audit from "../lib/audit.js";
-import type * as lib_issueTelemetry from "../lib/issueTelemetry.js";
-import type * as lib_timeRange from "../lib/timeRange.js";
 import type * as lib_backendIds from "../lib/backendIds.js";
 import type * as lib_backends_capabilities from "../lib/backends/capabilities.js";
 import type * as lib_backends_mock from "../lib/backends/mock.js";
@@ -54,6 +52,7 @@ import type * as lib_e2ee from "../lib/e2ee.js";
 import type * as lib_e2eeCrypto from "../lib/e2eeCrypto.js";
 import type * as lib_http from "../lib/http.js";
 import type * as lib_issuance from "../lib/issuance.js";
+import type * as lib_issueTelemetry from "../lib/issueTelemetry.js";
 import type * as lib_loadBands from "../lib/loadBands.js";
 import type * as lib_locations from "../lib/locations.js";
 import type * as lib_membershipCode from "../lib/membershipCode.js";
@@ -74,6 +73,7 @@ import type * as lib_statusCounters from "../lib/statusCounters.js";
 import type * as lib_statusPage from "../lib/statusPage.js";
 import type * as lib_supportId from "../lib/supportId.js";
 import type * as lib_themeConfig from "../lib/themeConfig.js";
+import type * as lib_timeRange from "../lib/timeRange.js";
 import type * as lib_umami from "../lib/umami.js";
 import type * as lib_urlSafety from "../lib/urlSafety.js";
 import type * as lib_verificationConfig from "../lib/verificationConfig.js";
@@ -133,8 +133,6 @@ declare const fullApi: ApiFromModules<{
   "lib/accountId": typeof lib_accountId;
   "lib/analyticsConfig": typeof lib_analyticsConfig;
   "lib/audit": typeof lib_audit;
-  "lib/issueTelemetry": typeof lib_issueTelemetry;
-  "lib/timeRange": typeof lib_timeRange;
   "lib/backendIds": typeof lib_backendIds;
   "lib/backends/capabilities": typeof lib_backends_capabilities;
   "lib/backends/mock": typeof lib_backends_mock;
@@ -153,6 +151,7 @@ declare const fullApi: ApiFromModules<{
   "lib/e2eeCrypto": typeof lib_e2eeCrypto;
   "lib/http": typeof lib_http;
   "lib/issuance": typeof lib_issuance;
+  "lib/issueTelemetry": typeof lib_issueTelemetry;
   "lib/loadBands": typeof lib_loadBands;
   "lib/locations": typeof lib_locations;
   "lib/membershipCode": typeof lib_membershipCode;
@@ -173,6 +172,7 @@ declare const fullApi: ApiFromModules<{
   "lib/statusPage": typeof lib_statusPage;
   "lib/supportId": typeof lib_supportId;
   "lib/themeConfig": typeof lib_themeConfig;
+  "lib/timeRange": typeof lib_timeRange;
   "lib/umami": typeof lib_umami;
   "lib/urlSafety": typeof lib_urlSafety;
   "lib/verificationConfig": typeof lib_verificationConfig;
