+++ messages/ar.json
@@ -742,111 +742,7 @@
       "how": "البداية",
       "membership": "العضوية",
       "faq": "الأسئلة الشائعة",
-      "about": "حول",
-      "globe": "الخريطة"
-    },
-    "globe": {
-      "title": "أصوات من خلف الجدار الناري",
-      "body": "كل يوم، يستخدم الناس خلف الجدران النارية الوطنية FreeSocks للقراءة والمشاركة والنشر — وليُسمع صوتهم.",
-      "aria": "كرة أرضية دوارة: ملصقات تحمل أصواتًا من دول خاضعة للرقابة",
-      "voices": {
-        "ir": {
-          "place": "طهران، إيران",
-          "l1": "يحجبون الإنترنت؛ لكنهم لا يستطيعون حجب الحقيقة.",
-          "l2": "تقريري عن الاحتجاجات وصل إلى العالم.",
-          "l3": "طلابي يقرأون ما تسميه الدولة أكاذيب."
-        },
-        "cn": {
-          "place": "بكين، الصين",
-          "l1": "لا ينبغي أن يكون التاريخ كلمة محظورة.",
-          "l2": "شاركت ما حدث في مدينتي — وبقي منشورًا.",
-          "l3": "أرشفت المنشورات المحذوفة قبل اختفائها."
-        },
-        "ru": {
-          "place": "موسكو، روسيا",
-          "l1": "الحقيقة لا ينبغي أن تحتاج إلى تصريح.",
-          "l2": "الصحافة المستقلة ليست جريمة.",
-          "l3": "الصحيفة المحجوبة ما زالت تُقرأ."
-        },
-        "tm": {
-          "place": "عشق آباد، تركمانستان",
-          "l1": "بلد كامل، شبه منقطع عن الإنترنت — ومع ذلك مسموع.",
-          "l2": "يتحكمون في الإعلام، لا في صوتي.",
-          "l3": "الصمت قانون هنا؛ ومع ذلك نتهامس."
-        },
-        "cu": {
-          "place": "هافانا، كوبا",
-          "l1": "صوتي يسافر أبعد مما سأصل إليه يومًا.",
-          "l2": "نوثق ما تنكره الدولة.",
-          "l3": "أصوات مستقلة، من يد إلى يد، ومن شاشة إلى شاشة."
-        },
-        "by": {
-          "place": "مينسك، بيلاروس",
-          "l1": "حين أوقفونا، واصلنا الكلام.",
-          "l2": "الانتخابات الحرة ليست تطرفًا.",
-          "l3": "حظروا علمنا؛ لا صوتنا."
-        },
-        "mm": {
-          "place": "يانغون، ميانمار",
-          "l1": "انقطاع الإنترنت لم يُسكتنا.",
-          "l2": "أدلة القمع خرجت إلى العلن.",
-          "l3": "حين صمتت الأبراج، لم تصمت الحكاية."
-        },
-        "ve": {
-          "place": "كاراكاس، فنزويلا",
-          "l1": "نحن نُحصي الأصوات التي لا يحصونها.",
-          "l2": "الجوع ليس سرًّا من أسرار الدولة.",
-          "l3": "طابور الخبز طويل؛ والحقيقة أطول."
-        },
-        "vn": {
-          "place": "هانوي، فيتنام",
-          "l1": "الكتابة عن الفساد ليست جريمة.",
-          "l2": "مدونتي نجت من الحجب.",
-          "l3": "مقال واحد حذفوه وصل إلى الآلاف."
-        },
-        "pk": {
-          "place": "كراتشي، باكستان",
-          "l1": "أثناء القطيعة، ظل الشهود يتكلمون.",
-          "l2": "إسكات الصحفيين لن يخفي القصة.",
-          "l3": "القناة أظلمت؛ الصحافة لم تُظلم."
-        },
-        "eg": {
-          "place": "القاهرة، مصر",
-          "l1": "سجنوا المدونين، لا الكلمات.",
-          "l2": "وُثّقت الاحتجاجات رغم كل شيء.",
-          "l3": "من شاشة صغيرة، قصة كبيرة."
-        },
-        "sa": {
-          "place": "الرياض، السعودية",
-          "l1": "الكلام ليس جريمة.",
-          "l2": "صوتها وصل إلى ما وراء الجدار.",
-          "l3": "سألت سؤالًا تحظره المملكة."
-        },
-        "et": {
-          "place": "أديس أبابا، إثيوبيا",
-          "l1": "القطيعة لم تُخفِ شيئًا.",
-          "l2": "الشهود وجدوا طريقًا للخارج.",
-          "l3": "أُحصي القتلى، رغم الانقطاع."
-        },
-        "tr": {
-          "place": "إسطنبول، تركيا",
-          "l1": "حجب الموقع لن يحجب القصة.",
-          "l2": "الصحافة مستمرة، بأمر المحكمة أو بدونه.",
-          "l3": "مذيع معتقل لا يستطيع إنهاء النشرة."
-        },
-        "az": {
-          "place": "باكو، أذربيجان",
-          "l1": "يسمون الصحافة تطرفًا.",
-          "l2": "نُشر التحقيق رغم كل شيء.",
-          "l3": "جمّدوا حساباتنا، لا عملنا."
-        },
-        "uz": {
-          "place": "طشقند، أوزبكستان",
-          "l1": "الإنترنت المغلق ليس إنترنت صامتًا.",
-          "l2": "وسمع العالم بما جرى هنا.",
-          "l3": "الساحات مراقبة؛ والكلمات تتحرك رغم ذلك."
-        }
-      }
+      "about": "حول"
     },
     "impact": {
       "title": "التبرعات تُحدث فرقًا",
