+++ messages/fa.json
@@ -742,111 +742,7 @@
       "how": "شروع به کار",
       "membership": "عضویت",
       "faq": "سوالات متداول",
-      "about": "درباره",
-      "globe": "نقشه"
-    },
-    "globe": {
-      "title": "صداهایی از پشت دیوار آتشین",
-      "body": "هر روز مردم پشت دیوارهای آتشین ملی با FreeSocks می‌خوانند، به اشتراک می‌گذارند، منتشر می‌کنند — و شنیده می‌شوند.",
-      "aria": "کره‌ای چرخان: برچسب‌هایی با صداهایی از کشورهای سانسورشده",
-      "voices": {
-        "ir": {
-          "place": "تهران، ایران",
-          "l1": "اینترنت را فیلتر می‌کنند؛ حقیقت را نمی‌توانند فیلتر کنند.",
-          "l2": "گزارش من دربارهٔ اعتراض‌ها به جهان رسید.",
-          "l3": "شاگردانم آنچه را دولت دروغ می‌خواند می‌خوانند."
-        },
-        "cn": {
-          "place": "پکن، چین",
-          "l1": "تاریخ نباید کلمه‌ای ممنوع باشد.",
-          "l2": "آنچه در شهرم گذشت را به اشتراک گذاشتم — و باقی ماند.",
-          "l3": "پست‌های حذف‌شده را پیش از نابودیشان بایگانی کردم."
-        },
-        "ru": {
-          "place": "مسکو، روسیه",
-          "l1": "حقیقت نباید به مجوز نیاز داشته باشد.",
-          "l2": "روزنامه‌نگاری مستقل جرم نیست.",
-          "l3": "روزنامهٔ مسدودشده همچنان خوانده می‌شود."
-        },
-        "tm": {
-          "place": "عشق‌آباد، ترکمنستان",
-          "l1": "یک کشور کامل، تقریباً آفلاین — و با این حال شنیده‌شده.",
-          "l2": "رسانه‌ها را کنترل می‌کنند، نه صدایم را.",
-          "l3": "سکوت این‌جا قانون است؛ ما با این حال پچ‌پچ می‌کنیم."
-        },
-        "cu": {
-          "place": "هاوانا، کوبا",
-          "l1": "صدای من خیلی دورتر از خودم می‌رسد.",
-          "l2": "آنچه دولت انکار می‌کند را ثبت می‌کنیم.",
-          "l3": "صداهای مستقل، دست به دست، صفحه به صفحه."
-        },
-        "by": {
-          "place": "مینسک، بلاروس",
-          "l1": "وقتی خاموشمان کردند، باز هم سخن گفتیم.",
-          "l2": "انتخابات آزاد افراطی‌گری نیست.",
-          "l3": "پرچممان را ممنوع کردند؛ صدایمان را نه."
-        },
-        "mm": {
-          "place": "یانگون، میانمار",
-          "l1": "قطعی اینترنت ساکتمان نکرد.",
-          "l2": "شواهد سرکوب بیرون رفت.",
-          "l3": "وقتی برج‌ها خاموش شدند، قصه خاموش نشد."
-        },
-        "ve": {
-          "place": "کاراکاس، ونزوئلا",
-          "l1": "رأی‌هایی را می‌شماریم که آن‌ها نمی‌شمارند.",
-          "l2": "گرسنگی راز دولتی نیست.",
-          "l3": "صف نان طولانی است؛ حقیقت طولانی‌تر."
-        },
-        "vn": {
-          "place": "هانوی، ویتنام",
-          "l1": "نوشتن دربارهٔ فساد جرم نیست.",
-          "l2": "وبلاگم از مسدودیت جان سالم به در برد.",
-          "l3": "همان یک مقاله‌ای که حذف کردند به هزاران نفر رسید."
-        },
-        "pk": {
-          "place": "کراچی، پاکستان",
-          "l1": "در زمان قطعی، شاهدان همچنان سخن گفتند.",
-          "l2": "ساکت‌کردن روزنامه‌نگاران خبر را پنهان نمی‌کند.",
-          "l3": "شبکه تاریک شد؛ خبرنگاری نشد."
-        },
-        "eg": {
-          "place": "قاهره، مصر",
-          "l1": "وبلاگ‌نویسان را زندانی کردند، نه کلمات را.",
-          "l2": "اعتراض به هر حال ثبت شد.",
-          "l3": "از یک صفحهٔ کوچک، قصه‌ای بزرگ."
-        },
-        "sa": {
-          "place": "ریاض، عربستان سعودی",
-          "l1": "حرف زدن جرم نیست.",
-          "l2": "صدای او از پس دیوار بیرون رفت.",
-          "l3": "او پرسشی پرسید که پادشاهی ممنوع می‌کند."
-        },
-        "et": {
-          "place": "آدیس‌آبابا، اتیوپی",
-          "l1": "قطعی اینترنت هیچ‌چیز را پنهان نکرد.",
-          "l2": "شاهدان باز هم راهی بیرون پیدا کردند.",
-          "l3": "کشته‌ها شمرده شدند، با وجود قطعی."
-        },
-        "tr": {
-          "place": "استانبول، ترکیه",
-          "l1": "مسدودکردن سایت، خبر را مسدود نمی‌کند.",
-          "l2": "روزنامه‌نگاری ادامه دارد، با حکم دادگاه هم.",
-          "l3": "مجری دستگیرشده نمی‌تواند خبر را تمام کند."
-        },
-        "az": {
-          "place": "باکو، آذربایجان",
-          "l1": "آن‌ها به گزارش‌گری می‌گویند افراطی‌گری.",
-          "l2": "تحقیق به هر حال منتشر شد.",
-          "l3": "حساب‌هایمان را مسدود کردند، نه کارمان را."
-        },
-        "uz": {
-          "place": "تاشکند، ازبکستان",
-          "l1": "اینترنت بسته، اینترنت ساکتی نیست.",
-          "l2": "جهان باز هم شنید که این‌جا چه گذشت.",
-          "l3": "میدان‌ها تحت نظرند؛ کلمات با این حال می‌گردند."
-        }
-      }
+      "about": "درباره"
     },
     "impact": {
       "title": "کمک‌های مالی در حال کار",
