+++ messages/ru.json
@@ -749,111 +749,7 @@
       "how": "С чего начать",
       "membership": "Членство",
       "faq": "ЧаВо",
-      "about": "О нас",
-      "globe": "Карта"
-    },
-    "globe": {
-      "title": "Голоса из-за файрвола",
-      "body": "Каждый день люди за государственными файрволами читают, делятся и публикуют через FreeSocks — и их слышат.",
-      "aria": "Вращающийся глобус: подписи с голосами из стран с цензурой",
-      "voices": {
-        "ir": {
-          "place": "Тегеран, Иран",
-          "l1": "Они фильтруют интернет, но правду не отфильтруешь.",
-          "l2": "Мой репортаж о протестах увидел весь мир.",
-          "l3": "Мои студенты читают то, что государство называет ложью."
-        },
-        "cn": {
-          "place": "Пекин, Китай",
-          "l1": "История не должна быть запрещённым словом.",
-          "l2": "Я рассказал, что случилось в моём городе, — и это осталось в сети.",
-          "l3": "Я заархивировал удалённые посты до того, как они исчезли."
-        },
-        "ru": {
-          "place": "Москва, Россия",
-          "l1": "Правда не должна требовать разрешения.",
-          "l2": "Независимая журналистика — не преступление.",
-          "l3": "Заблокированную газету всё равно читают."
-        },
-        "tm": {
-          "place": "Ашхабад, Туркменистан",
-          "l1": "Целая страна почти офлайн — и всё равно услышанная.",
-          "l2": "Они контролируют СМИ, но не мой голос.",
-          "l3": "Молчание здесь — закон; но мы всё равно шепчемся."
-        },
-        "cu": {
-          "place": "Гавана, Куба",
-          "l1": "Мой голос улетает дальше, чем я когда-либо смогу.",
-          "l2": "Мы документируем то, что государство отрицает.",
-          "l3": "Независимые голоса — из рук в руки, с экрана на экран."
-        },
-        "by": {
-          "place": "Минск, Беларусь",
-          "l1": "Когда нас отключали, мы всё равно говорили.",
-          "l2": "Свободные выборы — это не экстремизм.",
-          "l3": "Они запретили наш флаг, но не наш голос."
-        },
-        "mm": {
-          "place": "Янгон, Мьянма",
-          "l1": "Блэкаут не заставил нас замолчать.",
-          "l2": "Доказательства репрессий вышли наружу.",
-          "l3": "Когда вышки замолчали, история не замолчала."
-        },
-        "ve": {
-          "place": "Каракас, Венесуэла",
-          "l1": "Мы считаем голоса, которые они не считают.",
-          "l2": "Голод — не государственная тайна.",
-          "l3": "Очередь за хлебом длинная; правда длиннее."
-        },
-        "vn": {
-          "place": "Ханой, Вьетнам",
-          "l1": "Писать о коррупции — не преступление.",
-          "l2": "Мой блог пережил блокировку.",
-          "l3": "Одна удалённая статья дошла до тысяч."
-        },
-        "pk": {
-          "place": "Карачи, Пакистан",
-          "l1": "Во время отключения свидетели продолжали говорить.",
-          "l2": "Заткнув журналистов, историю не скроешь.",
-          "l3": "Канал погас; журналистика — нет."
-        },
-        "eg": {
-          "place": "Каир, Египет",
-          "l1": "Блогеров посадили, но не слова.",
-          "l2": "Протесты всё равно задокументировали.",
-          "l3": "С маленького экрана — большая история."
-        },
-        "sa": {
-          "place": "Эр-Рияд, Саудовская Аравия",
-          "l1": "Говорить — не преступление.",
-          "l2": "Её голос прорвался за стену.",
-          "l3": "Она задала вопрос, запрещённый в королевстве."
-        },
-        "et": {
-          "place": "Аддис-Абеба, Эфиопия",
-          "l1": "Отключение ничего не скрыло.",
-          "l2": "Свидетели всё равно нашли выход.",
-          "l3": "Погибших посчитали, несмотря на блэкаут."
-        },
-        "tr": {
-          "place": "Стамбул, Турция",
-          "l1": "Заблокировать сайт — не заблокировать историю.",
-          "l2": "Журналистика продолжается, с решением суда или без.",
-          "l3": "Арестованный ведущий не может закончить выпуск."
-        },
-        "az": {
-          "place": "Баку, Азербайджан",
-          "l1": "Они называют репортажи экстремизмом.",
-          "l2": "Расследование всё равно вышло.",
-          "l3": "Они заморозили наши счета, но не нашу работу."
-        },
-        "uz": {
-          "place": "Ташкент, Узбекистан",
-          "l1": "Закрытый интернет — не тихий интернет.",
-          "l2": "Мир всё равно услышал, что здесь произошло.",
-          "l3": "Площади под наблюдением; слова всё равно движутся."
-        }
-      }
+      "about": "О нас"
     },
     "impact": {
       "title": "Пожертвования в деле",
