+++ messages/zh.json
@@ -742,111 +742,7 @@
       "how": "快速上手",
       "membership": "会员",
       "faq": "常见问题",
-      "about": "关于",
-      "globe": "地图"
-    },
-    "globe": {
-      "title": "来自防火墙后的声音",
-      "body": "每天，身处国家防火墙后的人们使用 FreeSocks 阅读、分享和发布内容——并让世界听见。",
-      "aria": "旋转的地球仪：带有来自受审查国家声音的标签",
-      "voices": {
-        "ir": {
-          "place": "德黑兰，伊朗",
-          "l1": "他们过滤互联网，却过滤不了真相。",
-          "l2": "我关于抗议的报道传到了全世界。",
-          "l3": "我的学生读着被国家称为谎言的东西。"
-        },
-        "cn": {
-          "place": "北京，中国",
-          "l1": "历史不该成为违禁词。",
-          "l2": "我分享了家乡发生的事——它一直还在。",
-          "l3": "被删的帖子，我在它们消失前就存了档。"
-        },
-        "ru": {
-          "place": "莫斯科，俄罗斯",
-          "l1": "真相不该需要许可证。",
-          "l2": "独立新闻不是犯罪。",
-          "l3": "被封的报纸，依然有人读。"
-        },
-        "tm": {
-          "place": "阿什哈巴德，土库曼斯坦",
-          "l1": "整个国家几乎断网——却依然被听见。",
-          "l2": "他们控制媒体，却控制不了我的声音。",
-          "l3": "沉默是这里的法律；我们依然低声细语。"
-        },
-        "cu": {
-          "place": "哈瓦那，古巴",
-          "l1": "我的声音，比我走得更远。",
-          "l2": "我们记录国家否认的一切。",
-          "l3": "独立的声音，手手相传，屏屏相递。"
-        },
-        "by": {
-          "place": "明斯克，白俄罗斯",
-          "l1": "他们断网时，我们仍在发声。",
-          "l2": "自由选举不是极端主义。",
-          "l3": "他们禁了我们的旗帜，却禁不了我们的声音。"
-        },
-        "mm": {
-          "place": "仰光，缅甸",
-          "l1": "断网没能让我们沉默。",
-          "l2": "镇压的证据传了出去。",
-          "l3": "信号塔沉默了，故事却没有。"
-        },
-        "ve": {
-          "place": "加拉加斯，委内瑞拉",
-          "l1": "我们清点他们不肯清点的选票。",
-          "l2": "饥饿不是国家机密。",
-          "l3": "买面包的队伍很长；真相更长。"
-        },
-        "vn": {
-          "place": "河内，越南",
-          "l1": "揭露腐败不是犯罪。",
-          "l2": "我的博客，比封锁活得更久。",
-          "l3": "那篇被删的文章，传到了成千上万人手中。"
-        },
-        "pk": {
-          "place": "卡拉奇，巴基斯坦",
-          "l1": "断网期间，目击者仍在发声。",
-          "l2": "让记者沉默，掩盖不了真相。",
-          "l3": "频道黑了，报道没有停。"
-        },
-        "eg": {
-          "place": "开罗，埃及",
-          "l1": "他们关押了博主，却关不住文字。",
-          "l2": "抗议依然被记录了下来。",
-          "l3": "小小的屏幕，大大的故事。"
-        },
-        "sa": {
-          "place": "利雅得，沙特阿拉伯",
-          "l1": "说话不是犯罪。",
-          "l2": "她的声音传到了墙外。",
-          "l3": "她问了一个王国禁止的问题。"
-        },
-        "et": {
-          "place": "亚的斯亚贝巴，埃塞俄比亚",
-          "l1": "断网掩盖不了一切。",
-          "l2": "目击者还是把消息传了出去。",
-          "l3": "逝者被一一记录，尽管断网。"
-        },
-        "tr": {
-          "place": "伊斯坦布尔，土耳其",
-          "l1": "封了网站，封不住真相。",
-          "l2": "无论有没有法院禁令，新闻仍在继续。",
-          "l3": "被捕的主播，无法为新闻画上句号。"
-        },
-        "az": {
-          "place": "巴库，阿塞拜疆",
-          "l1": "他们把报道称为极端主义。",
-          "l2": "调查报道还是发表了。",
-          "l3": "他们冻结了我们的账户，却冻不住我们的工作。"
-        },
-        "uz": {
-          "place": "塔什干，乌兹别克斯坦",
-          "l1": "封闭的互联网，并不等于沉默的互联网。",
-          "l2": "世界依然听到了这里发生的事。",
-          "l3": "广场有人监视；文字仍在流动。"
-        }
-      }
+      "about": "关于"
     },
     "impact": {
       "title": "捐款在行动",
