+++ src/client/lib/i18n/message-keys.ts
@@ -396,71 +396,9 @@ export type MessageKey =
   | "home.freeCard.urlTitle"
   | "home.globe.aria"
   | "home.globe.body"
+  | "home.globe.legendRegions"
+  | "home.globe.legendServers"
   | "home.globe.title"
-  | "home.globe.voices.az.l1"
-  | "home.globe.voices.az.l2"
-  | "home.globe.voices.az.l3"
-  | "home.globe.voices.az.place"
-  | "home.globe.voices.by.l1"
-  | "home.globe.voices.by.l2"
-  | "home.globe.voices.by.l3"
-  | "home.globe.voices.by.place"
-  | "home.globe.voices.cn.l1"
-  | "home.globe.voices.cn.l2"
-  | "home.globe.voices.cn.l3"
-  | "home.globe.voices.cn.place"
-  | "home.globe.voices.cu.l1"
-  | "home.globe.voices.cu.l2"
-  | "home.globe.voices.cu.l3"
-  | "home.globe.voices.cu.place"
-  | "home.globe.voices.eg.l1"
-  | "home.globe.voices.eg.l2"
-  | "home.globe.voices.eg.l3"
-  | "home.globe.voices.eg.place"
-  | "home.globe.voices.et.l1"
-  | "home.globe.voices.et.l2"
-  | "home.globe.voices.et.l3"
-  | "home.globe.voices.et.place"
-  | "home.globe.voices.ir.l1"
-  | "home.globe.voices.ir.l2"
-  | "home.globe.voices.ir.l3"
-  | "home.globe.voices.ir.place"
-  | "home.globe.voices.mm.l1"
-  | "home.globe.voices.mm.l2"
-  | "home.globe.voices.mm.l3"
-  | "home.globe.voices.mm.place"
-  | "home.globe.voices.pk.l1"
-  | "home.globe.voices.pk.l2"
-  | "home.globe.voices.pk.l3"
-  | "home.globe.voices.pk.place"
-  | "home.globe.voices.ru.l1"
-  | "home.globe.voices.ru.l2"
-  | "home.globe.voices.ru.l3"
-  | "home.globe.voices.ru.place"
-  | "home.globe.voices.sa.l1"
-  | "home.globe.voices.sa.l2"
-  | "home.globe.voices.sa.l3"
-  | "home.globe.voices.sa.place"
-  | "home.globe.voices.tm.l1"
-  | "home.globe.voices.tm.l2"
-  | "home.globe.voices.tm.l3"
-  | "home.globe.voices.tm.place"
-  | "home.globe.voices.tr.l1"
-  | "home.globe.voices.tr.l2"
-  | "home.globe.voices.tr.l3"
-  | "home.globe.voices.tr.place"
-  | "home.globe.voices.uz.l1"
-  | "home.globe.voices.uz.l2"
-  | "home.globe.voices.uz.l3"
-  | "home.globe.voices.uz.place"
-  | "home.globe.voices.ve.l1"
-  | "home.globe.voices.ve.l2"
-  | "home.globe.voices.ve.l3"
-  | "home.globe.voices.ve.place"
-  | "home.globe.voices.vn.l1"
-  | "home.globe.voices.vn.l2"
-  | "home.globe.voices.vn.l3"
-  | "home.globe.voices.vn.place"
   | "home.hero.subtitle"
   | "home.hero.title"
   | "home.how.cta"
